UndefinedType

Description

This module is for exercise code which is incomplete module. In order to pass typechecking with incomplete module, use @Undefined data type.

Documentation

data Undefined

A dummy type

(===) :: (Eq a, Show a) => a -> a -> a infixl 0